+++ controlplane/src/core/bufservices/feature-flag/createFeatureFlag.ts
@@ -205,6 +205,7 @@ export function createFeatureFlag(
           cdnBaseUrl: opts.cdnBaseUrl,
           webhookJWTSecret: opts.admissionWebhookJWTSecret,
         },
+        chClient: opts.chClient!,
       });
 
       compositionErrors.push(...composition.compositionErrors);