+++ controlplane/CHANGELOG.md
@@ -4,6 +4,24 @@ Binaries are attached to the github release otherwise all images can be found [h
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.120.2](https://github.com/wundergraph/cosmo/compare/controlplane@0.120.1...controlplane@0.120.2) (2025-01-19)
+
+### Bug Fixes
+
+* **cache operation:** swallow cache errors and other improvements ([#1515](https://github.com/wundergraph/cosmo/issues/1515)) ([d959e2c](https://github.com/wundergraph/cosmo/commit/d959e2c9fb492cc7c73d89f61c31f3bad2ac5706)) (@StarpTech)
+
+## [0.120.1](https://github.com/wundergraph/cosmo/compare/controlplane@0.120.0...controlplane@0.120.1) (2025-01-18)
+
+### Bug Fixes
+
+* **cache warmup:** consider only po of the last 7 days ([#1513](https://github.com/wundergraph/cosmo/issues/1513)) ([1d2f041](https://github.com/wundergraph/cosmo/commit/1d2f0410c240eaeda97b2d7a8afafd4173b672f1)) (@StarpTech)
+
+# [0.120.0](https://github.com/wundergraph/cosmo/compare/controlplane@0.119.1...controlplane@0.120.0) (2025-01-17)
+
+### Features
+
+* cache warmer ([#1501](https://github.com/wundergraph/cosmo/issues/1501)) ([948edd2](https://github.com/wundergraph/cosmo/commit/948edd23e6d0ee968c91edd1a9e9943c3405ac2d)) (@JivusAyrus)
+
 ## [0.119.1](https://github.com/wundergraph/cosmo/compare/controlplane@0.119.0...controlplane@0.119.1) (2025-01-08)
 
 ### Bug Fixes